Ensure AMD dependencies are always in the expected place
npm will deduplicate dependencies by omitting them from children
and expecting that filesystem walking will occur. The loader could
be modified to perform this action, but when client.html is loaded
directly by an end-user, it would not be possible to do this. As
a result, we simply make sure that symlinks to dependencies loaded
via AMD on the client are always put in place. (Dependencies loaded
as Node.js modules already work properly.)